+++ src/it/scala/de/upb/cs/swt/delphi/webapi/SearchQueryTest.scala
@@ -16,7 +16,6 @@
 
 package de.upb.cs.swt.delphi.webapi
 
-import de.upb.cs.swt.delphi.webapi.featuredefinitions.FeatureExtractor
 import de.upb.cs.swt.delphi.webapi.search.{QueryRequest, SearchQuery}
 import org.scalatest.{FlatSpec, Matchers}
 
@@ -25,7 +24,7 @@ import scala.util.Success
 class SearchQueryTest extends FlatSpec with Matchers {
   "Search query" should "check for fields" in {
     val configuration = new Configuration()
-    val q = new SearchQuery(configuration, new FeatureExtractor(configuration))
+    val q = new SearchQuery(configuration, new FeatureQuery(configuration))
 
     val response = q.search(QueryRequest("[if_icmpeq (opcode:159)]>1"))
     response shouldBe a [Success[_]]
