--- src/main/scala/de/upb/cs/swt/delphi/webapi/featuredefinitions/FeatureListMapping.scala
@@ -1,421 +1,0 @@
-// Copyright (C) 2018 The Delphi Team.
-// See the LICENCE file distributed with this work for additional
-// information regarding copyright ownership.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package de.upb.cs.swt.delphi.webapi.featuredefinitions
-
-object FeatureListMapping {
-
-
-  //Returns a list of all defined features
-  def featureList: List[String] =
-    (for {(key, des) <- featureMap if des != null} yield key).toList
-
-  //Maps all field names onto descriptions of that field. Null fields cannot be searched in the current version
-  private val featureMap = Map[String, FeatureDescription](
-    "ProjectPackages" -> null,
-    "⟨SizeOfInheritanceTree⟩" -> null,
-    "ProjectFields" -> null,
-    "LibraryMethods" -> null,
-    "ProjectInstructions" -> null,
-    "ProjectMethods" -> null,
-    "LibraryClassFiles" -> null,
-    "LibraryFields" -> null,
-    "LibraryPackages" -> null,
-    "ProjectClassFiles" -> null,
-    "0 FPC" -> null,
-    "1-3 FPC" -> null,
-    "4-10 FPC" -> null,
-    ">10 FPC" -> null,
-    "0 MPC" -> null,
-    "1-3 MPC" -> null,
-    "4-10 MPC" -> null,
-    ">10 MPC" -> null,
-    "1-3 CPP" -> null,
-    "4-10 CPP" -> null,
-    ">10 CPP" -> null,
-    "0 NOC" -> null,
-    "1-3 NOC" -> null,
-    "4-10 NOC" -> null,
-    ">10 NOC" -> null,
-    "linear methods (McCabe)" -> null,
-    "2-3 McCabe" -> null,
-    "4-10 McCabe" -> null,
-    ">10 McCabe" -> null,
-    "Designator" -> null,
-    "Taxonomy" -> null,
-    "Joiner" -> null,
-    "Pool" -> null,
-    "Function Pointer" -> null,
-    "Function Object" -> null,
-    "Cobol Like" -> null,
-    "Stateless" -> null,
-    "Common State" -> null,
-    "Immutable" -> null,
-    "Restricted Creation" -> null,
-    "Sampler" -> null,
-    "Box" -> null,
-    "Compound Box" -> null,
-    "Canopy" -> null,
-    "Record" -> null,
-    "Data Manager" -> null,
-    "Sink" -> null,
-    "Outline" -> null,
-    "Trait" -> null,
-    "State Machine" -> null,
-    "Pure Type" -> null,
-    "Augmented Type" -> null,
-    "Pseudo Class" -> null,
-    "Implementor" -> null,
-    "Overrider" -> null,
-    "Extender" -> null,
-    "unused private fields" -> null,
-    "unused package visible fields" -> null,
-    "unused protected fields" -> null,
-    "unused public fields" -> null,
-    "package visible fields only used by defining type" -> null,
-    "protected fields only used by defining type" -> null,
-    "public fields only used by defininig type " -> null,
-    "Trivial Class.forName Usage" -> null,
-    "Nontrivial Class.forName Usage" -> null,
-    "nop (opcode:0)" -> null,
-    "aconst_null (opcode:1)" -> null,
-    "iconst_m1 (opcode:2)" -> null,
-    "iconst_0 (opcode:3)" -> null,
-    "iconst_1 (opcode:4)" -> null,
-    "iconst_2 (opcode:5)" -> null,
-    "iconst_3 (opcode:6)" -> null,
-    "iconst_4 (opcode:7)" -> null,
-    "iconst_5 (opcode:8)" -> null,
-    "lconst_0 (opcode:9)" -> null,
-    "lconst_1 (opcode:10)" -> null,
-    "fconst_0 (opcode:11)" -> null,
-    "fconst_1 (opcode:12)" -> null,
-    "fconst_2 (opcode:13)" -> null,
-    "dconst_0 (opcode:14)" -> null,
-    "dconst_1 (opcode:15)" -> null,
-    "bipush (opcode:16)" -> null,
-    "sipush (opcode:17)" -> null,
-    "ldc (opcode:18)" -> null,
-    "ldc_w (opcode:19)" -> null,
-    "ldc2_w (opcode:20)" -> null,
-    "iload (opcode:21)" -> null,
-    "lload (opcode:22)" -> null,
-    "fload (opcode:23)" -> null,
-    "dload (opcode:24)" -> null,
-    "aload (opcode:25)" -> null,
-    "iload_0 (opcode:26)" -> null,
-    "iload_1 (opcode:27)" -> null,
-    "iload_2 (opcode:28)" -> null,
-    "iload_3 (opcode:29)" -> null,
-    "lload_0 (opcode:30)" -> null,
-    "lload_1 (opcode:31)" -> null,
-    "lload_2 (opcode:32)" -> null,
-    "lload_3 (opcode:33)" -> null,
-    "fload_0 (opcode:34)" -> null,
-    "fload_1 (opcode:35)" -> null,
-    "fload_2 (opcode:36)" -> null,
-    "fload_3 (opcode:37)" -> null,
-    "dload_0 (opcode:38)" -> null,
-    "dload_1 (opcode:39)" -> null,
-    "dload_2 (opcode:40)" -> null,
-    "dload_3 (opcode:41)" -> null,
-    "aload_0 (opcode:42)" -> null,
-    "aload_1 (opcode:43)" -> null,
-    "aload_2 (opcode:44)" -> null,
-    "aload_3 (opcode:45)" -> null,
-    "iaload (opcode:46)" -> null,
-    "laload (opcode:47)" -> null,
-    "faload (opcode:48)" -> null,
-    "daload (opcode:49)" -> null,
-    "aaload (opcode:50)" -> null,
-    "baload (opcode:51)" -> null,
-    "caload (opcode:52)" -> null,
-    "saload (opcode:53)" -> null,
-    "istore (opcode:54)" -> null,
-    "lstore (opcode:55)" -> null,
-    "fstore (opcode:56)" -> null,
-    "dstore (opcode:57)" -> null,
-    "astore (opcode:58)" -> null,
-    "istore_0 (opcode:59)" -> null,
-    "istore_1 (opcode:60)" -> null,
-    "istore_2 (opcode:61)" -> null,
-    "istore_3 (opcode:62)" -> null,
-    "lstore_0 (opcode:63)" -> null,
-    "lstore_1 (opcode:64)" -> null,
-    "lstore_2 (opcode:65)" -> null,
-    "lstore_3 (opcode:66)" -> null,
-    "fstore_0 (opcode:67)" -> null,
-    "fstore_1 (opcode:68)" -> null,
-    "fstore_2 (opcode:69)" -> null,
-    "fstore_3 (opcode:70)" -> null,
-    "dstore_0 (opcode:71)" -> null,
-    "dstore_1 (opcode:72)" -> null,
-    "dstore_2 (opcode:73)" -> null,
-    "dstore_3 (opcode:74)" -> null,
-    "astore_0 (opcode:75)" -> null,
-    "astore_1 (opcode:76)" -> null,
-    "astore_2 (opcode:77)" -> null,
-    "astore_3 (opcode:78)" -> null,
-    "iastore (opcode:79)" -> null,
-    "lastore (opcode:80)" -> null,
-    "fastore (opcode:81)" -> null,
-    "dastore (opcode:82)" -> null,
-    "aastore (opcode:83)" -> null,
-    "bastore (opcode:84)" -> null,
-    "castore (opcode:85)" -> null,
-    "sastore (opcode:86)" -> null,
-    "pop (opcode:87)" -> null,
-    "pop2 (opcode:88)" -> null,
-    "dup (opcode:89)" -> null,
-    "dup_x1 (opcode:90)" -> null,
-    "dup_x2 (opcode:91)" -> null,
-    "dup2 (opcode:92)" -> null,
-    "dup2_x1 (opcode:93)" -> null,
-    "dup2_x2 (opcode:94)" -> null,
-    "swap (opcode:95)" -> null,
-    "iadd (opcode:96)" -> null,
-    "ladd (opcode:97)" -> null,
-    "fadd (opcode:98)" -> null,
-    "dadd (opcode:99)" -> null,
-    "isub (opcode:100)" -> null,
-    "lsub (opcode:101)" -> null,
-    "fsub (opcode:102)" -> null,
-    "dsub (opcode:103)" -> null,
-    "imul (opcode:104)" -> null,
-    "lmul (opcode:105)" -> null,
-    "fmul (opcode:106)" -> null,
-    "dmul (opcode:107)" -> null,
-    "idiv (opcode:108)" -> null,
-    "ldiv (opcode:109)" -> null,
-    "fdiv (opcode:110)" -> null,
-    "ddiv (opcode:111)" -> null,
-    "irem (opcode:112)" -> null,
-    "lrem (opcode:113)" -> null,
-    "frem (opcode:114)" -> null,
-    "drem (opcode:115)" -> null,
-    "ineg (opcode:116)" -> null,
-    "lneg (opcode:117)" -> null,
-    "fneg (opcode:118)" -> null,
-    "dneg (opcode:119)" -> null,
-    "ishl (opcode:120)" -> null,
-    "lshl (opcode:121)" -> null,
-    "ishr (opcode:122)" -> null,
-    "lshr (opcode:123)" -> null,
-    "iushr (opcode:124)" -> null,
-    "lushr (opcode:125)" -> null,
-    "iand (opcode:126)" -> null,
-    "land (opcode:127)" -> null,
-    "ior (opcode:128)" -> null,
-    "lor (opcode:129)" -> null,
-    "ixor (opcode:130)" -> null,
-    "lxor (opcode:131)" -> null,
-    "iinc (opcode:132)" -> null,
-    "i2l (opcode:133)" -> null,
-    "i2f (opcode:134)" -> null,
-    "i2d (opcode:135)" -> null,
-    "l2i (opcode:136)" -> null,
-    "l2f (opcode:137)" -> null,
-    "l2d (opcode:138)" -> null,
-    "f2i (opcode:139)" -> null,
-    "f2l (opcode:140)" -> null,
-    "f2d (opcode:141)" -> null,
-    "d2i (opcode:142)" -> null,
-    "d2l (opcode:143)" -> null,
-    "d2f (opcode:144)" -> null,
-    "i2b (opcode:145)" -> null,
-    "i2c (opcode:146)" -> null,
-    "i2s (opcode:147)" -> null,
-    "lcmp (opcode:148)" -> null,
-    "fcmpl (opcode:149)" -> null,
-    "fcmpg (opcode:150)" -> null,
-    "dcmpl (opcode:151)" -> null,
-    "dcmpg (opcode:152)" -> null,
-    "ifeq (opcode:153)" -> null,
-    "ifne (opcode:154)" -> null,
-    "iflt (opcode:155)" -> null,
-    "ifge (opcode:156)" -> null,
-    "ifgt (opcode:157)" -> null,
-    "ifle (opcode:158)" -> null,
-    "if_icmpeq (opcode:159)" -> null,
-    "if_icmpne (opcode:160)" -> null,
-    "if_icmplt (opcode:161)" -> null,
-    "if_icmpge (opcode:162)" -> null,
-    "if_icmpgt (opcode:163)" -> null,
-    "if_icmple (opcode:164)" -> null,
-    "if_acmpeq (opcode:165)" -> null,
-    "if_acmpne (opcode:166)" -> null,
-    "goto (opcode:167)" -> null,
-    "jsr (opcode:168)" -> null,
-    "ret (opcode:169)" -> null,
-    "tableswitch (opcode:170)" -> null,
-    "lookupswitch (opcode:171)" -> null,
-    "ireturn (opcode:172)" -> null,
-    "lreturn (opcode:173)" -> null,
-    "freturn (opcode:174)" -> null,
-    "dreturn (opcode:175)" -> null,
-    "areturn (opcode:176)" -> null,
-    "return (opcode:177)" -> null,
-    "getstatic (opcode:178)" -> null,
-    "putstatic (opcode:179)" -> null,
-    "getfield (opcode:180)" -> null,
-    "putfield (opcode:181)" -> null,
-    "invokevirtual (opcode:182)" -> null,
-    "invokespecial (opcode:183)" -> null,
-    "invokestatic (opcode:184)" -> null,
-    "invokeinterface (opcode:185)" -> null,
-    "invokedynamic (opcode:186)" -> null,
-    "new (opcode:187)" -> null,
-    "newarray (opcode:188)" -> null,
-    "anewarray (opcode:189)" -> null,
-    "arraylength (opcode:190)" -> null,
-    "athrow (opcode:191)" -> null,
-    "checkcast (opcode:192)" -> null,
-    "instanceof (opcode:193)" -> null,
-    "monitorenter (opcode:194)" -> null,
-    "monitorexit (opcode:195)" -> null,
-    "wide (opcode:196)" -> null,
-    "multianewarray (opcode:197)" -> null,
-    "ifnull (opcode:198)" -> null,
-    "ifnonnull (opcode:199)" -> null,
-    "goto_w (opcode:200)" -> null,
-    "jsr_w (opcode:201)" -> null,
-    "Self-recursive Data Structure" -> null,
-    "Mutually-recursive Data Structure 2 Types" -> null,
-    "Mutually-recursive Data Structure 3 Types" -> null,
-    "Mutually-recursive Data Structure 4 Types" -> null,
-    "Mutually-recursive Data Structure more than 4 Types" -> null,
-    "Never Returns Normally" -> null,
-    "Method with Infinite Loop" -> null,
-    "Class File With Source Attribute" -> null,
-    "Method With Line Number Table" -> null,
-    "Method With Local Variable Table" -> null,
-    "Method With Local Variable Type Table" -> null,
-    "FanOut - Category 1" -> null,
-    "FanOut - Category 2" -> null,
-    "FanOut - Category 3" -> null,
-    "FanOut - Category 4" -> null,
-    "FanOut - Category 5" -> null,
-    "FanOut - Category 6" -> null,
-    "FanIn - Category 1" -> null,
-    "FanIn - Category 2" -> null,
-    "FanIn - Category 3" -> null,
-    "FanIn - Category 4" -> null,
-    "FanIn - Category 5" -> null,
-    "FanIn - Category 6" -> null,
-    "FanIn/FanOut - Category 1" -> null,
-    "FanIn/FanOut - Category 2" -> null,
-    "FanIn/FanOut - Category 3" -> null,
-    "FanIn/FanOut - Category 4" -> null,
-    "FanIn/FanOut - Category 5" -> null,
-    "FanIn/FanOut - Category 6" -> null,
-    "custom ClassLoader implementation" -> null,
-    "Retrieving the SystemClassLoader" -> null,
-    "Retrieving some ClassLoader" -> null,
-    "define new classes/packages" -> null,
-    "accessing resources" -> null,
-    "javax.crypto.Cipher getInstance" -> null,
-    "using SecureRandom" -> null,
-    "using MessageDigest" -> null,
-    "using Signature" -> null,
-    "using Mac" -> null,
-    "cryptographic key handling" -> null,
-    "using KeyStore" -> null,
-    "using Certificates" -> null,
-    "native methods" -> null,
-    "synthetic methods" -> null,
-    "bridge methods" -> null,
-    "synchronized methods" -> null,
-    "varargs methods" -> null,
-    "static initializers" -> null,
-    "static methods (not including static initializers)" -> null,
-    "constructors" -> null,
-    "instance methods" -> null,
-    "java.lang.Class forName" -> null,
-    "reflective instance creation" -> null,
-    "reflective field write" -> null,
-    "reflective field read" -> null,
-    "makes fields accessible" -> null,
-    "makes methods or constructors accessible" -> null,
-    "makes an AccessibleObject accessible (exact type unknown)" -> null,
-    "java.lang.reflect.Method Object invoke(Object, Object[])" -> null,
-    "java.lang.invoke.MethodHandles lookup" -> null,
-    "java.lang.invoke.MethodHandles publicLookup" -> null,
-    "method handle invocation" -> null,
-    "java.lang.reflect.Proxy newProxyInstance" -> null,
-    "Process" -> null,
-    "JVM exit" -> null,
-    "Native Libraries" -> null,
-    "java.lang.System getSecurityManager" -> null,
-    "java.lang.System setSecurityManager" -> null,
-    "Environment" -> null,
-    "Sound" -> null,
-    "Network sockets" -> null,
-    "Object-based Thread Notification" -> null,
-    "Usage of Thread API" -> null,
-    "Usage of ThreadGroup API" -> null,
-    "sun.misc.Unsafe sun.misc.Unsafe getUnsafe()" -> null,
-    "Unsafe - Alloc" -> null,
-    "Unsafe - Array" -> null,
-    "Unsafe - compareAndSwap" -> null,
-    "Unsafe - Class" -> null,
-    "Unsafe - Fence" -> null,
-    "Unsafe - Fetch & Add" -> null,
-    "Unsafe - Heap" -> null,
-    "Unsafe - Heap Get" -> null,
-    "Unsafe - Heap Put" -> null,
-    "Misc" -> null,
-    "Unsafe - Monitor" -> null,
-    "Unsafe - Off-Heap" -> null,
-    "Unsafe - Offset" -> null,
-    "Unsafe - Ordered Put" -> null,
-    "Unsafe - Park" -> null,
-    "Unsafe - Throw" -> null,
-    "Unsafe - Volatile Get" -> null,
-    "Unsafe - Volatile Put" -> null,
-    "java.sql.DriverManager getConnection" -> null,
-    "java.sql.Connection rollback" -> null,
-    "creation and execution of java.sql.Statement" -> null,
-    "creation and execution of java.sql.PreparedStatement" -> null,
-    "creation and execution of java.sql.CallableStatement" -> null,
-    "class file retransformation" -> null,
-    "instrumenting native methods" -> null,
-    "appending class loader search" -> null,
-    "retrieve classes information" -> null,
-    "(concrete) classes" -> null,
-    "abstract classes" -> null,
-    "annotations" -> null,
-    "enumerations" -> null,
-    "marker interfaces" -> null,
-    "simple functional interfaces (single abstract method (SAM) interface)" -> null,
-    "non-functional interface with default methods (Java >8)" -> null,
-    "non-functional interface with static methods (Java >8)" -> null,
-    "(standard) interface" -> null,
-    "module (Java >9)" -> null,
-    "Very Small Inheritance Tree" -> null,
-    "Small Inheritance Tree" -> null,
-    "Medium Inheritance Tree" -> null,
-    "Large Inheritance Tree" -> null,
-    "Very Large Inheritance Tree" -> null,
-    "Huge Inheritance Tree" -> null,
-    "Size of the Inheritance Tree Unknown" -> null,
-    "Class File JDK 1.1 (JDK 1.0.2)" -> null,
-    "Class File Java 5" -> null,
-    "Class File Java 6" -> null,
-    "Class File Java 7" -> null,
-    "Class File Java 8" -> null,
-    "Class File Java 9" -> null
-  )
-}
